+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+### Added
+
+- Command _east build_ will now after every build step copy
+  `compile_commands.json`, if found, from the build directory to the project
+  directory. This makes job of locating this file easier for clangd. Help
+  description for _east build_ was updated to reflect that.
+
 ### Changed
 
 - Make east.yml optional for everything, except for the usage of east build